refactor: ProjectDetailView のコンポーネント分離
- page.tsx を Thin page 化（データ取得を View に移動）
- VulnerabilitySection を Client Component として分離
- AiReportButton を独立コンポーネントとして抽出
- ProjectDetailView を Server Component としてデータ取得を担当
